Constructed the network
Constructed the GraphConvolutionalNetwork with the use of the GraphConvolutionalBlock.
In the forward in the input og the GraphBlock a single tensor of the batch was used. For this reason in the test is inserted the loop on the batch to obtain a batch of dimension 1 and in the forward is used the first index to obtain the tensor withouth the batch